refactor(uploads): move the reference-data writers out of the sqlpacketload route
upsertAttributeRows and upsertSpeciesRows are database writers living inside a
request handler, and nothing outside the route could reach them: a Next.js route
module may only export route fields, so adding `export` there fails the build
with `"upsertAttributeRows" is not a valid Route export field`.

Moves them, verbatim, to lib/uploads/reference-data-writers.ts along with the
helpers they own (normalizeOptionalString, normalizeRequiredString,
findDuplicateSpeciesCodes, truncateAndJoin, formatBlockedCleanReuploadValues,
MAX_LISTED_ISSUES, FixedDataProcessingResult). The route imports the four
symbols it still uses; upsertPersonnelRows stays put and keeps working through
the imported normalizers.

No logic change. Verified by sqlpacketload-authz plus the full integration run.
